app/src/test/java/com/appsdev/testslearn/ui/ShoppingViewModelTest.kt
package com.appsdev.testslearn.ui

import androidx.arch.core.executor.testing.InstantTaskExecutorRule
import com.appsdev.testslearn.MainCoroutineRule
import com.appsdev.testslearn.getOrAwaitValueTest
import com.appsdev.testslearn.other.Constants
import com.appsdev.testslearn.other.Status
import com.appsdev.testslearn.repositories.FakeShoppingRepository
import com.google.common.truth.Truth.assertThat
import kotlinx.coroutines.ExperimentalCoroutinesApi
import org.junit.Before
import org.junit.Rule
import org.junit.Test

@ExperimentalCoroutinesApi
class ShoppingViewModelTest {

    @get:Rule var instantTaskExecutorRule = InstantTaskExecutorRule()
    @get:Rule var mainCoroutineRule = MainCoroutineRule()
    private lateinit var viewModel: ShoppingViewModel

    @Before
    fun setUp() {
        viewModel = ShoppingViewModel(repository = FakeShoppingRepository())
    }

    @Test
    fun insertWithEmptyField_returnFalse() {
        viewModel.insert("name", "", "2.0")
        val value = viewModel.insertShoppingItemStatus.getOrAwaitValueTest()
        assertThat(value.getContentIfNotHandled()?.status).isEqualTo(Status.ERROR)
    }

    @Test
    fun insertWith2LongNames_returnFalse() {
        val string = buildString {
            for (i in 1..Constants.MAX_NAME_LENGTH + 1) {
                append(i)
            }
        }

        viewModel.insert(string, "11", "2.0")
        val value = viewModel.insertShoppingItemStatus.getOrAwaitValueTest()
        assertThat(value.getContentIfNotHandled()?.status).isEqualTo(Status.ERROR)
    }

    @Test
    fun insertWithTooLongPrice_returnFalse() {
        val string = buildString {
            for (i in 1..Constants.MAX_PRICE_LENGTH + 1) {
                append(i)
            }
        }

        viewModel.insert("name", "11", string)
        val value = viewModel.insertShoppingItemStatus.getOrAwaitValueTest()
        assertThat(value.getContentIfNotHandled()?.status).isEqualTo(Status.ERROR)
    }

    @Test
    fun insertWithInvalidAmount_returnFalse() {
        viewModel.insert("name", "999999999999999999999999999", "3.0")
        val value = viewModel.insertShoppingItemStatus.getOrAwaitValueTest()
        assertThat(value.getContentIfNotHandled()?.status).isEqualTo(Status.ERROR)
    }

    @Test
    fun insertWithValidInput_returnTrue() {
        viewModel.insert("name", "5", "3.0")
        val value = viewModel.insertShoppingItemStatus.getOrAwaitValueTest()
        assertThat(value.getContentIfNotHandled()?.status).isEqualTo(Status.SUCCESS)
    }
}